合入 master Phase0–5 与剩余项（恢复门禁、备份队列、预取 single-flight、
后台分页/翻页收敛、黑暗模式与夜读 chrome、死代码清理等），同时维持离线
产品线边界：

- 删除 AI 模块 / HttpTTS / AIHarness；设置无 AI 配置行
- 剥离 PageVC 翻译、Backup AI 条目、远程规则 URL 拉取
- run_all_harnesses 改跑 OfflineOnly；版本 1.8.0
- OfflineOnlyHarness 全绿；模拟器 Debug 构建通过

+++ Reader/Reader/Common/LocalBook/RDZipArchive.m
@@ -394,7 +394,8 @@ static void appendU32(NSMutableData *data, uint32_t value) {
     @catch (NSException *exception) {
         return NO;
     }
-    self.offset += header.length + data.length;
+    // Bounds already checked above (non-zip64); cast summed NSUInteger after guard.
+    self.offset += (uint32_t)(header.length + data.length);
     return YES;
 }
 
@@ -490,7 +491,8 @@ static void appendU32(NSMutableData *data, uint32_t value) {
         return NO;
     }
     [self.records addObject:record];
-    self.offset += header.length + record.size;
+    // Bounds already checked above (non-zip64); cast summed size after guard.
+    self.offset += (uint32_t)(header.length + record.size);
     return YES;
 }
